% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllCcsPredFun.R
\name{CcsPredict}
\alias{CcsPredict}
\title{CcsPredict}
\usage{
CcsPredict(mol_smiles, mol_names, base_dir = ".", thread = 3,
  is_output = TRUE)
}
\arguments{
\item{mol_smiles}{a vector of SMILES structures}

\item{mol_names}{a vector of identifier}

\item{base_dir}{'.'}

\item{thread}{Default: 3}

\item{is_output}{Default: TRUE}
}
\description{
A function to predict CCS values for small molecules with the input of SMILES structures
}
\examples{
test <- read.csv(system.file("extdata", "demo_data.csv", package="AllCCS"), stringsAsFactors = F)
test <- MdGet(mol_smiles = test$smiles, mol_names = test$id_allccs, base_dir = '.')
}
\author{
Zhiwei Zhou
\email{zhouzw@sioc.ac.cn}
}
