% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllCcsTools.R
\name{ExactMassCalculate}
\alias{ExactMassCalculate}
\title{ExactMassCalculate}
\usage{
ExactMassCalculate(formula = "C2H5OH")
}
\arguments{
\item{formula}{chemical formula}
}
\description{
Calculate exact mass using formula
}
\examples{
ExactMassCalculate("C2H5OH")
}
\author{
Zhiwei Zhou
\email{zhouzw@sioc.ac.cn}
}
