% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllCcsRssFun.R
\name{FpCalculate}
\alias{FpCalculate}
\title{FpCalcualte}
\usage{
FpCalculate(mol_smiles, mol_names, type = "pubchem", thread = 2,
  base_dir = ".")
}
\arguments{
\item{thread}{Default: 2}

\item{base_dir}{'.'}
}
\description{
FpCalcualte
}
\examples{
result_fps <- FpCalculate(mol_smiles = mol_smiles,
                          mol_names = mol_names,
                          thread = 2)
}
\author{
Zhiwei Zhou
}
