% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllCcsAnnotFun.R
\name{GetCcsRange}
\alias{GetCcsRange}
\title{GetCcsRange}
\usage{
GetCcsRange(data, type = c("percentage", "a_square"), tolerance = 4)
}
\arguments{
\item{data}{Numeric.}

\item{type}{'percentage', 'a_square'. Default: percentage}

\item{tolerance}{Numeric. CCS deviation tolerance. Default: 4}
}
\value{
a table of range. column 1: minimum value; column 2: maxmium value.
}
\description{
GetCcsRange
}
\examples{
GetCcsRange(data = c(100, 200),
            type = 'percentage',
            tolerance_ccs = 4)
GetCcsRange(data = c(100, 200),
            type = 'a_square',
            tolerance_ccs = 4)
}
\author{
Zhiwei Zhou
}
