% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllCcsTools.R
\name{KnnImpute}
\alias{KnnImpute}
\title{KnnImpute}
\usage{
KnnImpute(raw_matrix, label, k = 10, rowmax = 0.5, colmax = 0.8,
  maxp = 1500, rng.seed = 362436069)
}
\arguments{
\item{raw_matrix}{row: feature; column: sample}

\item{label}{feature id}
}
\description{
impute NA using KNN algorthem
}
\author{
Zhiwei Zhou
}
