% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllCcsMdFun.R
\name{MdCalculate}
\alias{MdCalculate}
\title{MdCalculate}
\usage{
MdCalculate(mol_smiles, mol_names)
}
\arguments{
\item{mol_smiles}{a smiles structure of molecule}

\item{mol_names}{the name of structure}
}
\description{
MdCalculate
}
\examples{
MdCalculate(mol_smiles = "COC1=CC=C(CC(O)=O)C=C1", mol_names = "4-Methoxyphenylacetic Acid")
}
\author{
Zhiwei Zhou
}
