% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllCcsMdFun.R
\name{MdGet}
\alias{MdGet}
\title{MdGet}
\usage{
MdGet(mol_smiles, mol_names, status_label, thread = 2, base_dir = ".")
}
\arguments{
\item{mol_smiles}{a vector of smiles}

\item{mol_names}{a vector of names}

\item{thread}{the number of calculating cores. Default: 2}

\item{base_dir}{'.'}
}
\value{
a list of MDs. "md_pos": Molecular descriptors in positive mode; "md_neg": Molecular descriptors in negative mode
}
\description{
Calculate molecular descriptors (MD), and perform MD cleaning (i.e. m/z transformation, NA imputation & scaling)
}
\examples{
test <- read.csv(system.file("extdata", "demo_data.csv", package="AllCCS"), stringsAsFactors = F)
}
\author{
Zhiwei Zhou
\email{zhouzw@sioc.ac.cn}
}
