% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllCcsTools.R
\name{MzTransform}
\alias{MzTransform}
\title{MzTransform}
\usage{
MzTransform(M = NULL, adduct = NULL, formula = NULL,
  polarity = c("pos", "neg"))
}
\arguments{
\item{M}{a vector of exact mass.}

\item{adduct}{a vector of adducts.}

\item{formula}{chemical formula}

\item{polarity}{'pos' or 'neg'}
}
\description{
calculate the m/z of adducts
}
\examples{
mz_transform(M = 123.2324, adduct = c('[M+H]+', '[M+Na]+'), polarity = 'pos')
mz_transform(formula="C2H5OH", adduct = c('M-', '[M-H]-'), polarity = 'neg')
}
\author{
Zhiwei Zhou
\email{zhouzw@sioc.ac.cn}
\code{mz_transform(M, adduct, polarity=c('pos', 'neg'))}
}
