% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllCcsMdFun.R
\name{NaImputate}
\alias{NaImputate}
\title{NaImpute}
\usage{
NaImputate(result_md_trans, polarity = c("pos", "neg"))
}
\arguments{
\item{polarity}{Default: pos}
}
\description{
NaImpute
}
\author{
Zhiwei Zhou
}
