% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllCcsRssFun.R
\name{RssCalculate}
\alias{RssCalculate}
\title{RssCalculate}
\usage{
RssCalculate(mol_smiles, mol_names, thread = 2, max_n = 5,
  polarity = c("pos", "neg"), type = "pubchem", method = "tanimoto",
  base_dir = ".")
}
\arguments{
\item{mol_smiles}{a vector of SMILES structures}

\item{mol_names}{a vector of identifier}

\item{thread}{Default: 2}

\item{max_n}{Top n for calculate RSS score}

\item{polarity}{'pos', 'neg'}

\item{type}{The type of fingerprint. See \code{\link{get.fingerprint}}.}

\item{method}{The type of distance metric to use. See \code{\link{fp.sim.matrix}}.}
}
\value{
a vector of rss score
}
\description{
Calculate representative structure similarity
}
\examples{
test <- RssCalculate(mol_smiles = mol_smiles, mol_names = mol_names, thread = 2, max_n = 5, polarity = 'pos')
}
\author{
Zhiwei Zhou
\email{zhouzw@sioc.ac.cn}
}
