% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllCcsTools.R
\name{Smiles2Mass}
\alias{Smiles2Mass}
\title{Smiles2Mass}
\usage{
Smiles2Mass(mol_smiles, mol_names)
}
\arguments{
\item{mol_smiles}{See \code{\link{get.mol2formula}}}
}
\value{
a vector of mass
}
\description{
Smiles2Mass
}
\examples{
test_smiles <- c("COC1=CC=C(CC(O)=O)C=C1", "CN(C)CCC1=CC=C(O)C=C1", "OC(=O)CCCCCCCC\\\\C=C\\\\C(O)=O")
test_name <- seq(length(test))
Smiles2Mass(mol_smiles = test, mol_names = test_name)
}
\author{
Zhiwei Zhou
}
