% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllCcsTools.R
\name{SmilesCheck}
\alias{SmilesCheck}
\title{SmilesCheck}
\usage{
SmilesCheck(mol_smiles)
}
\arguments{
\item{mol_smiles}{See \code{\link{is.smiles}}}
}
\value{
a vector of status, valid or error 1
}
\description{
SmilesCheck
}
\author{
Zhiwei Zhou
}
