% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllCcsMdFun.R
\name{ZscoreScaling}
\alias{ZscoreScaling}
\title{ZscoreScaling}
\usage{
ZscoreScaling(result_md_trans, polarity = c("pos", "neg"))
}
\arguments{
\item{result_md_trans}{Mds after mz transformation & NA imputation}
}
\value{
a data.frame of MDs for new structures
}
\description{
z-score transformation for MDs of new structure
}
\examples{

}
\author{
Zhiwei Zhou
}
