% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllCcsAnnotFun.R
\name{CcsAnnotate}
\alias{CcsAnnotate}
\title{CcsAnnotate}
\usage{
CcsAnnotate(mz = NULL, ccs = NULL, candidate_list = NULL,
  rank_type = c("ccs_scoring", "ccs_filtering"),
  filter_type = c("percentage", "a_square"), tolerance_ccs = 4,
  tolerance_min = 2, tolerance_max = 4, weight_ccs = 0.3,
  weight_msms = 0.7, base_dir = ".", thread = 1, is_output = TRUE)
}
\arguments{
\item{base_dir}{Default: '.'}

\item{is_output}{TRUE}
}
\description{
CcsAnnotate
}
\examples{
# CCS scoring
candidate_list <- system.file("extdata", "demo_annotation.csv", package="AllCCS")
test <- CcsAnnotate(mz = 666.312,
                    ccs = 261.0582,
                    candidate_list = candidate_list,
                    rank_type = 'ccs_scoring',
                    tolerance_min = 2,
                    tolerance_max = 4,
                    weight_ccs = 0.3,
                    weight_msms = 0.7,
                    base_dir = '.',
                    is_output = TRUE)

# CCS filtering
test <- CcsAnnotate(mz = 666.312,
                    ccs = 261.0582,
                    candidate_list = candidate_list,
                    rank_type = 'ccs_filtering',
                    filter_type = 'percentage',
                    tolerance_ccs = 4,
                    base_dir = '.',
                    is_output = TRUE)
}
\author{
Zhiwei Zhou
}
